worker/pyproject.toml
[project]
name = "agent-hub-worker"
version = "0.1.0"
description = "SQS worker for agent-hub"
requires-python = ">=3.11"
dependencies = [
  "agent-hub-core",
  "boto3>=1.35.0",
  "google-api-python-client>=2.120",
  "google-auth>=2.28",
  "google-auth-httplib2>=0.2",
  "httpx>=0.27.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
]

[tool.uv.sources]
agent-hub-core = { workspace = true }

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

# Flat layout: this directory is the `worker` package root (worker.__main__, worker.main, worker.core, …)
[tool.setuptools.package-dir]
worker = "."
"worker.core" = "core"
"worker.queue" = "queue"
"worker.handlers" = "handlers"
"worker.handlers.aws" = "handlers/aws"
"worker.messaging" = "messaging"

[tool.setuptools]
packages = [
  "worker",
  "worker.core",
  "worker.queue",
  "worker.handlers",
  "worker.handlers.aws",
  "worker.messaging",
]
